Fix keyboard crash and icon not showing
- Color(toULong()) → Color(toInt()): the ULong constructor expects
  64-bit internal encoding, not 32-bit ARGB. This caused wrong/
  transparent colors and likely crashed Compose rendering.
- Add try-catch fallback in onCreateInputView to prevent silent crash.
- Remove leftover drawable/ic_launcher_foreground.xml that was
  overriding the PNG mipmap icons.

+++ ime-core/src/main/kotlin/space/manus/nacre/ime/keyboard/KeyboardScreen.kt
@@ -19,8 +19,8 @@ fun KeyboardScreen(service: NacreInputMethodService) {
     val layerManager = service.layerManager
     val layout = layerManager.currentLayout()
     val theme = ThemeProvider.loadSelectedTheme(service)
-    val bgColor = Color(theme.background.toULong())
-    val accentColor = Color(theme.accent.toULong())
+    val bgColor = Color(theme.background.toInt())
+    val accentColor = Color(theme.accent.toInt())
 
     // Panel state
     var showClipboard by remember { mutableStateOf(false) }
